Break the program up into proper files.
